requirements-dev.txt
/home/shurjo/projects/propraetor/requirements-dev.txt#L1-3
# Development dependencies for Propraetor
django-debug-toolbar>=4.3,<5.0